backend/requirements.txt
openai==0.27.8
tenacity==8.2.2
python-dotenv==1.0.0